Make RenderResult an interface to enable RenderTargets for different UIs
